Refactor to group the models
